pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.124.0",
    "google-genai>=1.0.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
]
